rename consumer-group-not-found.ts → consumer-group-helpers.ts
